Minor map update

+++ code/game/machinery/OpTable.dm
@@ -41,14 +41,9 @@
 		else
 	return
 
-/obj/machinery/optable/blob_act()
-	if(prob(75))
-		qdel(src)
-
 /obj/machinery/optable/attack_hand(mob/user as mob)
 	if (HULK in usr.mutations)
-		usr << text("\blue You destroy the table.")
-		visible_message("\red [usr] destroys the operating table!")
+		visible_message("<span class='danger'>\The [usr] destroys \the [src]!</span>")
 		src.density = 0
 		qdel(src)
 	return
@@ -87,9 +82,9 @@
 
 /obj/machinery/optable/proc/take_victim(mob/living/carbon/C, mob/living/carbon/user as mob)
 	if (C == user)
-		user.visible_message("[user] climbs on the operating table.","You climb on the operating table.")
+		user.visible_message("[user] climbs on \the [src].","You climb on \the [src].")
 	else
-		visible_message("\red [C] has been laid on the operating table by [user].", 3)
+		visible_message("<span class='notice'>\The [C] has been laid on \the [src] by [user].</span>", 3)
 	if (C.client)
 		C.client.perspective = EYE_PERSPECTIVE
 		C.client.eye = src
@@ -108,10 +103,26 @@
 /obj/machinery/optable/MouseDrop_T(mob/target, mob/user)
 
 	var/mob/living/M = user
-	if(user.stat || user.restrained() || !check_table(user) || !iscarbon(target))
+	if(user.stat || user.restrained() ||  !iscarbon(target))
 		return
 	if(istype(M))
-		take_victim(target,user)
+
+		var/mob/living/L = target
+		var/bucklestatus = L.bucklecheck(user)
+
+		if (!bucklestatus)//We must make sure the person is unbuckled before they go in
+			return
+
+
+		if(L == user)
+			visible_message("[user] starts climbing onto the operating table.", 3)
+		else
+			visible_message("[user] starts putting [L.name] onto the operating table.", 3)
+		if (do_mob(user, L, 10, needhand = 0))
+			if (bucklestatus == 2)
+				var/obj/structure/LB = L.buckled
+				LB.user_unbuckle_mob(user)
+			take_victim(target,user)
 	else
 		return ..()
 
@@ -120,7 +131,7 @@
 	set category = "Object"
 	set src in oview(1)
 
-	if(usr.stat || !ishuman(usr) || usr.restrained() || !check_table(usr))
+	if(usr.stat || !ishuman(usr) || usr.restrained() )
 		return
 
 	take_victim(usr,usr)
@@ -128,18 +139,35 @@
 /obj/machinery/optable/attackby(obj/item/weapon/W as obj, mob/living/carbon/user as mob)
 	if (istype(W, /obj/item/weapon/grab))
 		var/obj/item/weapon/grab/G = W
-		if(iscarbon(G.affecting) && check_table(G.affecting))
+		if (src.victim)
+			usr << "<span class='notice'><B>The table is already occupied!</B></span>"
+			return 0
+
+		var/mob/living/L = G.affecting
+		var/bucklestatus = L.bucklecheck(user)
+
+		if (!bucklestatus)//We must make sure the person is unbuckled before they go in
+			return
+
+
+		if(L == user)
+			visible_message("[user] starts climbing onto the operating table.", 3)
+		else
+			visible_message("[user] starts putting [L.name] onto the operating table.", 3)
+		if (do_mob(user, L, 10, needhand = 0))
+			if (bucklestatus == 2)
+				var/obj/structure/LB = L.buckled
+				LB.user_unbuckle_mob(user)
 			take_victim(G.affecting,usr)
 			qdel(W)
 			return
 
 /obj/machinery/optable/proc/check_table(mob/living/carbon/patient as mob)
-	if(src.victim)
-		usr << "\blue <B>The table is already occupied!</B>"
+	check_victim()
+	if(src.victim && get_turf(victim) == get_turf(src) && victim.lying)
+		usr << "<span class='warning'>\The [src] is already occupied!</span>"
 		return 0
-
 	if(patient.buckled)
-		usr << "\blue <B>Unbuckle first!</B>"
+		usr << "<span class='notice'>Unbuckle \the [patient] first!</span>"
 		return 0
-
 	return 1
